+++ tests/dates/test_BetweenDatesTransformer.py
@@ -1,9 +1,11 @@
 import datetime
 
 import narwhals as nw
+import numpy as np
 import pandas as pd
 import polars as pl
 import pytest
+from beartype.roar import BeartypeCallHintParamViolation
 from dateutil.tz import gettz
 
 import tests.test_data as d
@@ -33,6 +35,27 @@ class TestInit(
     @classmethod
     def setup_class(cls):
         cls.transformer_name = "BetweenDatesTransformer"
+
+    # overload until we beartype the new_column_name mixin
+    @pytest.mark.parametrize(
+        "new_column_type",
+        [1, True, {"a": 1}, [1, 2], np.inf, np.nan],
+    )
+    def test_new_column_name_type_error(
+        self,
+        new_column_type,
+        minimal_attribute_dict,
+        uninitialized_transformers,
+    ):
+        """Test an error is raised if any type other than str passed to new_column_name"""
+
+        args = minimal_attribute_dict[self.transformer_name].copy()
+        args["new_column_name"] = new_column_type
+
+        with pytest.raises(
+            BeartypeCallHintParamViolation,
+        ):
+            uninitialized_transformers[self.transformer_name](**args)
 
     @pytest.mark.parametrize(
         ("param", "value"),