+++ tubular/numeric.py
@@ -29,7 +29,7 @@ from tubular.mixins import (
     NewColumnNameMixin,
     TwoColumnMixin,
 )
-from tubular.types import DataFrame
+from tubular.types import DataFrame, ListOfOneStr, PositiveNumber
 
 if TYPE_CHECKING:
     from narwhals.typing import FrameT, IntoSeriesT
@@ -243,7 +243,10 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
     @beartype
     def __init__(
         self,
-        columns: Union[str, list[str]],
+        columns: Union[
+            str,
+            ListOfOneStr,
+        ],
         new_column_name: str,
         n_init: Union[str, int] = "auto",
         n_clusters: int = 8,
@@ -251,10 +254,6 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
         kmeans_kwargs: Optional[dict[str, object]] = None,
         **kwargs: dict[str, bool],
     ) -> None:
-        if (isinstance(columns, list)) and (len(columns) > 1):
-            msg = f"{self.classname()}: columns arg should be a single str or a list length 1 giving the column to group."
-            raise TypeError(msg)
-
         if kmeans_kwargs is None:
             kmeans_kwargs = {}
 
@@ -262,6 +261,7 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
         self.new_column_name = new_column_name
         self.n_init = n_init
         self.kmeans_kwargs = kmeans_kwargs
+        self.drop_original = drop_original
 
         if isinstance(columns, str):
             self.columns = [columns]
@@ -269,7 +269,6 @@ class OneDKmeansTransformer(BaseNumericTransformer, DropOriginalMixin):
             self.columns = columns
 
         super().__init__(columns=[columns], **kwargs)
-        self.set_drop_original_column(drop_original)
 
     def get_feature_names_out(self) -> list[str]:
         """list features modified/created by the transformer
@@ -502,10 +501,11 @@ class LogTransformer(BaseNumericTransformer, DropOriginalMixin):
 
     FITS = False
 
+    @beartype
     def __init__(
         self,
-        columns: str | list[str] | None,
-        base: float | None = None,
+        columns: Optional[Union[str, list[str]]],
+        base: Optional[PositiveNumber] = None,
         add_1: bool = False,
         drop_original: bool = True,
         suffix: str = "log",
@@ -513,24 +513,7 @@ class LogTransformer(BaseNumericTransformer, DropOriginalMixin):
     ) -> None:
         super().__init__(columns=columns, **kwargs)
 
-        if not isinstance(add_1, bool):
-            add_1_error_msg = f"{self.classname()}: add_1 should be bool"
-            raise TypeError(add_1_error_msg)
-
-        if not isinstance(suffix, str):
-            suffix_error_msg = f"{self.classname()}: suffix should be str"
-            raise TypeError(suffix_error_msg)
-
-        if base is not None:
-            if not isinstance(base, (int, float)):
-                msg = f"{self.classname()}: base should be numeric or None"
-                raise ValueError(msg)
-            if not base > 0:
-                msg = f"{self.classname()}: base should be strictly positive"
-                raise ValueError(msg)
-
-        self.set_drop_original_column(drop_original)
-
+        self.drop_original = drop_original
         self.base = base
         self.add_1 = add_1
         self.suffix = suffix